requirements.txt
amqp==2.2.2
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.3
celery==4.1.0
decorator==4.1.2
Django==1.11
django-celery-beat==1.1.0
django-redis==4.8.0
djangorestframework==3.7.7
enum34==1.1.6
ephem==3.7.6.0
ipython==5.5.0
ipython-genutils==0.2.0
kombu==4.1.0
mysqlclient==1.3.12
pathlib2==2.3.0
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.3
redis==2.10.6
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
vine==1.1.4
wcwidth==0.1.7
